+++ src/main/java/de/samply/document/DocumentType.java
@@ -4,5 +4,6 @@ public enum DocumentType {
     VOTUM,
     SCRIPT,
     PUBLICATION,
+    DESCRIPTION,
     OTHERS
 }